requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
discord==2.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
jproperties==2.1.1
multidict==6.0.5
numpy==2.0.0
opencv-python==4.10.0.84
pillow==10.4.0
portalocker==3.0.0
redis==5.2.1
six==1.16.0
yarl==1.9.4
python-dotenv==1.21.1
